backend/airflow/requirements.txt
apache-airflow==2.9.1
biopython
requests